backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
yfinance==0.2.28
openai==0.28.1
python-dotenv==1.0.0
sqlite3
lxml==4.9.3
html5lib==1.1